feature-processing/pyproject.toml
[tool.poetry]
name = "feature-processing"
version = "0.1.0"
description = ""
authors = ["dewasahu2003 <dewasahu2003@gmail.com>"]
readme = "README.md"
packages = [{include = "feature_processing"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic-settings = "^2.1.0"
pika = "^1.3.2"
qdrant-client = "^1.8.0"
langchain = "^0.1.13"
structlog = "^24.1.0"
bytewax = "0.19.0"
unstructured = "^0.14.2"
langchain-openai = "^0.1.3"
sentence-transformers = "^2.2.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
